@@ -1,24 +1,14 @@
-import webpack from 'webpack'
-import StaticSiteGeneratorPlugin from 'static-site-generator-webpack-plugin'
-import ExtractTextPlugin from 'extract-text-webpack-plugin'
-import Config from 'webpack-configurator'
-const debug = require('debug')('gatsby:webpack-config')
+import fs from 'fs'
 import path from 'path'
-import _ from 'lodash'
-import invariant from 'invariant'
+import webpack from 'webpack'
+import Config from 'webpack-configurator'
+import ExtractTextPlugin from 'extract-text-webpack-plugin'
+import StaticSiteGeneratorPlugin from 'static-site-generator-webpack-plugin'
 
 import babelConfig from './babel-config'
+import validateWebpackConfig from './validate-webpack-config'
 
-let modifyWebpackConfig
-try {
-  const gatsbyNodeConfig = path.resolve(process.cwd(), './gatsby-node')
-  const nodeConfig = require(gatsbyNodeConfig)
-  modifyWebpackConfig = nodeConfig.modifyWebpackConfig
-} catch (e) {
-  if (e.code !== 'MODULE_NOT_FOUND' && !_.includes(e.Error, 'gatsby-node')) {
-    console.log(e)
-  }
-}
+const debug = require('debug')('gatsby:webpack-config')
 
 // Five stages or modes:
 //   1) develop: for `gatsby develop` command, hot reload and CSS injection into page
@@ -416,6 +406,31 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
     }
   }
 
+  function resolveLoader () {
+    const root = [
+      path.resolve(__dirname, '..', 'loaders'),
+    ]
+
+    const userLoaderDirectoryPath = path.resolve(directory, 'loaders')
+
+    try {
+      if (fs.statSync(userLoaderDirectoryPath).isDirectory()) {
+        root.push(userLoaderDirectoryPath)
+      }
+    } catch (e) {
+      if (e && e.code !== 'ENOENT') {
+        console.log(e)
+      }
+    }
+
+    return {
+      root,
+      modulesDirectories: [
+        'node_modules',
+      ],
+    }
+  }
+
   const config = new Config()
 
   config.merge({
@@ -427,29 +442,10 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
     debug: true,
     devtool: devtool(),
     output: output(),
-    resolveLoader: {
-      root: [
-        path.resolve(directory, 'loaders'),
-        path.resolve(__dirname, '..', 'loaders'),
-      ],
-      modulesDirectories: [
-        'node_modules',
-      ],
-    },
+    resolveLoader: resolveLoader(),
     plugins: plugins(),
     resolve: resolve(),
   })
 
-  if (modifyWebpackConfig) {
-    const modifiedWebpackConfig = modifyWebpackConfig(module(config), stage)
-    invariant(_.isObject(modifiedWebpackConfig),
-              `
-              You must return an object when modifying the Webpack config.
-              Returned: ${modifiedWebpackConfig}
-              stage: ${stage}
-              `)
-    return modifiedWebpackConfig
-  } else {
-    return module(config)
-  }
+  return validateWebpackConfig(module, config, stage)
 }
