@@ -16,7 +16,7 @@ var verDigit = Number(versionDigits)
 
 if (verDigit < 0.12) {
   console.error(
-    'Error: Gatsby 0.9+ requires node.js v0.12 or higher (you have ' + version + ') ' +
+    'Error: Gatsby 0.13+ requires node.js v4 or higher (you have ' + version + ') ' +
     'Upgrade node to the latest stable release.'
   )
   process.exit()
