@@ -1,7 +1,9 @@
 /* @flow weak */
-import webpack from 'webpack'
-import webpackConfig from './webpack.config'
+
 import fs from 'fs'
+import webpack from 'webpack'
+
+import webpackConfig from './webpack.config'
 
 module.exports = (program, callback) => {
   const { directory } = program
